introduced a separate file name handler
This class is given the parameters at the start, and then when you want the file name you retreive it by specify the file type and it will automatically select the correct data and produce a name with a relative or absolute path.
Cleans up some of the code in the meg_bids file.
Also added the RecordingDuration property, and modified the file copying code a little (it never used to work for me...)