Convert row value to data dtype before comparison
`tsv_handler._contains_rows()` is called when writing the
participants TSV file. Since I have a few participants with
an unknown birthday, the respective entries in the TSV
are 'n/a'. As `_contains_rows()` converts the TSV data to
an `ndarray`, and an `ndarray` can only contain elements
of the same dtype, the resulting array would have dtype
`U3` (unicode) in that case.

Now, when constructing the `column_mask`, `np.in1d()`
is called: first argument is the above-mentioned ndarray,
and second argument is the current value in the row. In
the case of a birthday, to stay with that specific example,
it may happen that the "current value" is an int, say, 30,
but the ndarray we're feeding as the first argument to
`np.in1d()` is of dtype `U3` if any previously processed
participant didn't have a birthday set. So `np.in1d()` now
has to deal with a numpy dtype and a (different) pure-Python
dtype. Consequently, it spits out a `FutureWarning`:

```
FutureWarning: elementwise comparison failed; returning scalar,
but in the future will perform elementwise comparison
```

So this is something that might even break our code at some
point in the future.

My approach to fix this is the following:

- we first retrieve the dtype from the data ndarray
- and then cast the "current value" into an ndarray of
  the same dtype

  This successfully avoids the warning.