ENH: Allow concatenation of BIDSPath with  string
In the mne-study-template we constantly create derivatives, which are currently
not or not sufficiently specified in the BIDS standard. Hence, the way
we derive their filenames is typically something like:

```
basename = make_bids_basename(...)
fname =  op.join(bids_basename + '_processing_step.fif')
```
Yet, this concatenation doesn't work with BIDSPath anymore. Instead,
we'd have to convert the BIDSPath to a string first, which seems
cumbersome.

This commit therefore adds an `__add__()` method to BIDSPath, which will
allow for the above code to work unchanged.