ENH: Ignore entities from `derivatives` subfolder
When calling `get_entity_vals()`,  we now ignore the `derivatives`
folder in the BIDS root directory. This is important as the
derivatives may, for example, contain "mock" subjects like
"sub-grand_average" or similar things, which we are typically
not interested in when looking at the BIDS root. If the users wants
to process the data in `derivatives/`, they can just set the
`bids_root` kwarg to `bids_root/derivatives`, and everything will work
as expected.