BF: Filename extension must be inferred from fname
`get_head_mri_trans()` in `master` tries to infer the filename extension
from a BIDS basename; however, the basename, as is created inside that
function, doesn't have a suffix. Correctly, the extension has to be
inferred from `bids_fname`. This is what we used to do before GH-446
got merged. This commit restores the old behavior.

(I realized something was going wrong when MNE-BIDS started looking
 for some .pdf files, while I only have .fif. So beyond what I'm
 fixing here, there seems to be another bug in `parse_ext()`, causing
 it to report `.pdf` when really it shouldn't.)