Fix bug if no event descriptions are present; support stim_type column in events.tsv (#680)
* Fix bug if no event descriptions are present; support stim_type

If no event descriptions could be generated, we'd crash as we'd try
to access descriptions like an array, while our placeholder would only
be a string. This is fixed now.

The other part of this PR is related to the naming of the column in the
events.tsv file that contains the event description. According to
BIDS, this column is optional and shall be named `trial_type`. Now
I've come across multiple (older?) datasets where the column is
called `stim_type` instead. I've now added support for reading this
column if `stim_type` is not present. A warning will be issued in
this case, informing the user that the column name should be changed.

This was necessary to read the ds000117 datset from
https://openneuro.org/datasets/ds000117

* Update changelog

* Add author