Fix scans.tsv reading (#684)
I cannot fully reproduce this, but @dnacombo was experiencing issues
when trying to read a raw file, as reading of `scans.tsv` failed. Reason
was that NumPy's `loadtxt` couldn't cope with the `BIDSPath`. Passing
a `pathlib.Path` instead fixed the problem.

Like I said, I could not reproduce this problem locally, but passing a
`Path` instead of a `BIDSPath` is probably a good idea anyway.