Datatype handling when in write_raw_bids (#774)
* Datatype handling when in write_raw_bids

The user must now specify the datatype when calling write_raw_bids. Not automatic datatype inferring is done anymore in write_raw_bids. However, it is checked if the specified datatype is present in the dataset.

* Small bug and style fixe

* Fix Error

data type specifciation by user is only required if multiple data types are present in data

* datatype inferring for meg restored

* Move datatype check into _handle_datatype

* Apply suggestions from code review

Co-authored-by: Alexandre Gramfort <alexandre.gramfort@m4x.org>

* PEP fix

* Fix test_handle_datatype

* restore data type inferring for multiple types

* Update tests

* Fix handling of combined MEG-i/EEG data

* Apply suggestions from code review

Co-authored-by: Alexandre Gramfort <alexandre.gramfort@m4x.org>

* Review

Co-authored-by: Alexandre Gramfort <alexandre.gramfort@m4x.org>