Wipe out all measurement date info when reading scans.tsv
If we find the acquisition date in `scans.tsv`, we use it to replace
`raw.info['meas_date']` in `main`.

However, the acquisition time might be stored in a few other places too,
which can lead to issues when trying to write the file again. For example,
I updated `scans.tsv` of one of my datasets so MNE could load it properly.
But when trying to write the data again, MNE raised an exception, because
`raw.info['meas_id']['secs']` was out of range.

This commit ensures that these problems are a thing of the past 💪