Update docstring of get_entity_vals() "root" param
We found that get_entity_vals() can be a horrible performance bottleneck
on large datasets stored on a network-attached storage system. Luckily
for us, we figured that, instead of passing the BIDS root as the "root"
parameter, it was totally valid and functional to pass a sub-directory
(in this specific case, sub-X for individual subjects) instead, leading
to a speedup by one order of magnitude. However, this behavior was
so far undocumented. This commit changes the docstring to "officially"
expose this functionality.