Make get_head_mri_trans() respect datatype and suffix attributes
get_head_mri_trans() currently blindly replaces the datatype and suffix
of the provided MEG BIDSPath with 'meg'. However, when working with
derivative data, the suffix might be different, causing the function to
fail as it will create a BIDSPath that points to a non-existent file.

The change here makes it such that if the provided MEG BIDSPath has
a suffix and / or datatype attribute set, they will not be modified.
Only if the attributes are not set, we will fall back to 'meg'.