Make write_raw_bids() raise if an unsuitable extension is provided in the BIDSPath
Some extensions / formats are only allowed for certain data types, e.g.
.vhdr is (i)EEG-only.

When passing a BIDSPath containing both datatype and extension
to write_raw_bids(), in case of a mismatch we would previously simply
silently replace the incorrect extension, effectively writing to a
location (and file format) the user didn't expect.

We now raise an exception in such situations.

Fixes #1041