FIX: Pin sphinx-gallery
in MNE-BIDS doc/conf.py we have:

```
sphinx_gallery_conf["within_subsection_order"] = "mne_bids.utils._example_sorter"
```

But AFAICT this functionality wasn't workign prior to https://github.com/sphinx-gallery/sphinx-gallery/pull/1391 , which was included in the release of sphinx-agallery v0.19 (released Feb 12 2025).

https://sphinx-gallery.github.io/stable/changes.html#v0-19-0

Given that v0.19 was released less than 1 year ago, I think it is reasonable to assume that devs/contributors might have a sphinx-gallery version < 0.19 installed locally, in which case building the mne-bids docuemntation will fail, but the reason/solution won't be explicit.

So I think that it is best to be explicit about the minimum required sphinx-gallery version we require to be able to successfully build our docs.